Handle errors when adding repo

+++ lib/silo/commands/repo_add.rb
@@ -27,7 +27,9 @@
 require_relative '../command'
 require_relative '../silo'
 require_relative '../type'
+
 require 'yaml'
+require 'open3'
 require 'tty-prompt'
 
 module FlightSilo
@@ -35,9 +37,9 @@ module FlightSilo
     class RepoAdd < Command
       def run
         answers = prompt.collect do
-          type_name = key("type").ask("Provider type name:")
-          type = Type[type_name]
-          type.questions.each do |question|
+          types = Type.all.map { |t| [t.description, t.name] }.to_h
+          type = key("type").select("Provider type:", types)
+          Type[type].questions.each do |question|
             key(question[:id]).ask(question[:text]) do |q|
               q.required question[:validation][:required]
               if question[:validation].to_h.key?(:format)
@@ -48,14 +50,22 @@ module FlightSilo
           end
         end
         
-        puts answers.inspect
-
-        type = Type[answers["type"]]
+        type_name = answers["type"]
         puts "Obtaining silo details for '#{answers["name"]}'..."
 
-        type_dir = "#{Config.root}/etc/types/#{type.name}"
+        type_dir = "#{Config.root}/etc/types/#{type_name}"
         ENV["flight_SILO_types"] = "#{Config.root}/etc/types"
-        `/bin/bash #{type_dir}/actions/pull.sh #{answers["name"]} /cloud_metadata.yaml #{type_dir}/cloud_metadata.yaml #{answers["region"]} #{answers["access_key"]} #{answers["secret_key"]}`
+
+        stdout_str, stderr_str, status = Open3.capture3(
+          "/bin/bash #{type_dir}/actions/pull.sh #{answers["name"]} /cloud_metadata.yaml #{type_dir}/cloud_metadata.yaml #{answers["region"]} #{answers["access_key"]} #{answers["secret_key"]}"
+        )
+
+        unless status.success?
+          raise <<~OUT
+          Error validating credentials:
+          #{stderr_str.chomp}
+          OUT
+        end
 
         cloud_md = YAML.load_file("#{type_dir}/cloud_metadata.yaml")
         `rm "#{type_dir}/cloud_metadata.yaml"`
